Revert "add sequence wise batchnorm"
