+++ stt_layer_gru.py
@@ -2,7 +2,7 @@ from collections import namedtuple
 
 import mxnet as mx
 
-from batchnorm import batchnorm
+from stt_layer_batchnorm import batchnorm
 
 GRUState = namedtuple("GRUState", ["h"])
 GRUParam = namedtuple("GRUParam", ["gates_i2h_weight", "gates_i2h_bias",
@@ -15,15 +15,15 @@ GRUModel = namedtuple("GRUModel", ["rnn_exec", "symbol",
                                    "param_blocks"])
 
 
-def gru(num_hidden, indata, prev_state, param, seqidx, layeridx, dropout_rate=0., is_batchnorm=False, gamma=None, beta=None, moving_mean=None, moving_var=None, name=None):
+def gru(num_hidden, indata, prev_state, param, seqidx, layeridx, dropout=0., is_batchnorm=False, gamma=None, beta=None, name=None):
     """
     GRU Cell symbol
     Reference:
     * Chung, Junyoung, et al. "Empirical evaluation of gated recurrent neural
         networks on sequence modeling." arXiv preprint arXiv:1412.3555 (2014).
     """
-    if dropout_rate > 0.:
-        indata = mx.sym.Dropout(data=indata, p=dropout_rate)
+    if dropout > 0.:
+        indata = mx.sym.Dropout(data=indata, p=dropout)
     i2h = mx.sym.FullyConnected(data=indata,
                                 weight=param.gates_i2h_weight,
                                 bias=param.gates_i2h_bias,
@@ -32,10 +32,9 @@ def gru(num_hidden, indata, prev_state, param, seqidx, layeridx, dropout_rate=0.
 
     if is_batchnorm:
         if name is not None:
-            i2h = batchnorm(net=i2h, gamma=gamma, beta=beta, moving_mean=moving_mean, moving_var=moving_var,
-                            name="%s_batchnorm" % name)
+            i2h = batchnorm(net=i2h, gamma=gamma, beta=beta, name="%s_batchnorm" % name)
         else:
-            i2h = batchnorm(net=i2h, gamma=gamma, beta=beta, moving_mean=moving_mean, moving_var=moving_var)
+            i2h = batchnorm(net=i2h, gamma=gamma, beta=beta)
     h2h = mx.sym.FullyConnected(data=prev_state.h,
                                 weight=param.gates_h2h_weight,
                                 bias=param.gates_h2h_bias,
@@ -64,34 +63,35 @@ def gru(num_hidden, indata, prev_state, param, seqidx, layeridx, dropout_rate=0.
     return GRUState(h=next_h)
 
 
-def gru_unroll(net, num_gru_layer, seq_len, num_hidden_gru_list, dropout_rate=0., is_batchnorm=False, prefix="",
-               direction="forward"):
+def gru_unroll(net, num_gru_layer, seq_len,  num_hidden_gru_list, dropout=0., is_batchnorm=False, prefix="",
+               direction="forward", is_bucketing=False):
     if num_gru_layer > 0:
         param_cells = []
         last_states = []
         for i in range(num_gru_layer):
-            param_cells.append(GRUParam(gates_i2h_weight=mx.sym.Variable(prefix+"l%d_i2h_gates_weight" % i),
-                                        gates_i2h_bias=mx.sym.Variable(prefix+"l%d_i2h_gates_bias" % i),
-                                        gates_h2h_weight=mx.sym.Variable(prefix+"l%d_h2h_gates_weight" % i),
-                                        gates_h2h_bias=mx.sym.Variable(prefix+"l%d_h2h_gates_bias" % i),
-                                        trans_i2h_weight=mx.sym.Variable(prefix+"l%d_i2h_trans_weight" % i),
-                                        trans_i2h_bias=mx.sym.Variable(prefix+"l%d_i2h_trans_bias" % i),
-                                        trans_h2h_weight=mx.sym.Variable(prefix+"l%d_h2h_trans_weight" % i),
-                                        trans_h2h_bias=mx.sym.Variable(prefix+"l%d_h2h_trans_bias" % i)))
-            state = GRUState(h=mx.sym.Variable(prefix+"l%d_init_h" % i))
+            param_cells.append(GRUParam(gates_i2h_weight=mx.sym.Variable(prefix + "l%d_i2h_gates_weight" % i),
+                                        gates_i2h_bias=mx.sym.Variable(prefix + "l%d_i2h_gates_bias" % i),
+                                        gates_h2h_weight=mx.sym.Variable(prefix + "l%d_h2h_gates_weight" % i),
+                                        gates_h2h_bias=mx.sym.Variable(prefix + "l%d_h2h_gates_bias" % i),
+                                        trans_i2h_weight=mx.sym.Variable(prefix + "l%d_i2h_trans_weight" % i),
+                                        trans_i2h_bias=mx.sym.Variable(prefix + "l%d_i2h_trans_bias" % i),
+                                        trans_h2h_weight=mx.sym.Variable(prefix + "l%d_h2h_trans_weight" % i),
+                                        trans_h2h_bias=mx.sym.Variable(prefix + "l%d_h2h_trans_bias" % i)))
+            state = GRUState(h=mx.sym.Variable(prefix + "l%d_init_h" % i))
             last_states.append(state)
-        assert(len(last_states) == num_gru_layer)
-        # Sequence-wise batch norm (supports variable length in inference)
+        assert (len(last_states) == num_gru_layer)
+        # declare batchnorm param(gamma,beta) in timestep wise
         if is_batchnorm:
             batchnorm_gamma = []
             batchnorm_beta = []
-            batchnorm_moving_mean = []
-            batchnorm_moving_var = []
-            for l in range(num_gru_layer):
-                batchnorm_gamma.append(mx.sym.Variable(prefix + "l%d_i2h_gamma" % l))
-                batchnorm_beta.append(mx.sym.Variable(prefix + "l%d_i2h_beta" % l))
-                batchnorm_moving_mean.append(mx.sym.Variable(prefix + "l%d_batchnorm_moving_mean" % l))
-                batchnorm_moving_var.append(mx.sym.Variable(prefix + "l%d_batchnorm_moving_var" % l))
+            if is_bucketing:
+                for l in range(num_gru_layer):
+                    batchnorm_gamma.append(mx.sym.Variable(prefix + "l%d_i2h_gamma" % l))
+                    batchnorm_beta.append(mx.sym.Variable(prefix + "l%d_i2h_beta" % l))
+            else:
+                for seqidx in range(seq_len):
+                    batchnorm_gamma.append(mx.sym.Variable(prefix + "t%d_i2h_gamma" % seqidx))
+                    batchnorm_beta.append(mx.sym.Variable(prefix + "t%d_i2h_beta" % seqidx))
 
         hidden_all = []
         for seqidx in range(seq_len):
@@ -106,31 +106,43 @@ def gru_unroll(net, num_gru_layer, seq_len, num_hidden_gru_list, dropout_rate=0.
 
             # stack GRU
             for i in range(num_gru_layer):
-                if dropout_rate > 0.:
-                    hidden = mx.sym.Dropout(data=hidden, p=dropout_rate)
+                if i == 0:
+                    dp_ratio = 0.
+                else:
+                    dp_ratio = dropout
                 if is_batchnorm:
-                    next_state = gru(num_hidden_gru_list[i], indata=hidden,
-                                     prev_state=last_states[i],
-                                     param=param_cells[i],
-                                     seqidx=k, layeridx=i, dropout_rate=dropout_rate,
-                                     is_batchnorm=is_batchnorm,
-                                     gamma=batchnorm_gamma[i],
-                                     beta=batchnorm_beta[i],
-                                     moving_mean=batchnorm_moving_mean[i],
-                                     moving_var=batchnorm_moving_var[i],
-                                     name=prefix + ("l%d" % i)
-                                     )
+                    if is_bucketing:
+                        next_state = gru(num_hidden_gru_list[i], indata=hidden,
+                                         prev_state=last_states[i],
+                                         param=param_cells[i],
+                                         seqidx=k, layeridx=i, dropout=dp_ratio,
+                                         is_batchnorm=is_batchnorm,
+                                         gamma=batchnorm_gamma[i],
+                                         beta=batchnorm_beta[i],
+                                         name=prefix + ("t%d_l%d" % (seqidx, i))
+                                         )
+                    else:
+                        next_state = gru(num_hidden_gru_list[i], indata=hidden,
+                                         prev_state=last_states[i],
+                                         param=param_cells[i],
+                                         seqidx=k, layeridx=i, dropout=dp_ratio,
+                                         is_batchnorm=is_batchnorm,
+                                         gamma=batchnorm_gamma[k],
+                                         beta=batchnorm_beta[k],
+                                         name=prefix + ("t%d_l%d" % (seqidx, i))
+                                         )
                 else:
                     next_state = gru(num_hidden_gru_list[i], indata=hidden,
                                      prev_state=last_states[i],
                                      param=param_cells[i],
-                                     seqidx=k, layeridx=i, dropout_rate=dropout_rate,
-                                     is_batchnorm=is_batchnorm)
+                                     seqidx=k, layeridx=i, dropout=dp_ratio,
+                                     is_batchnorm=is_batchnorm,
+                                     name=prefix)
                 hidden = next_state.h
                 last_states[i] = next_state
             # decoder
-            # if dropout_rate > 0.:
-            #     hidden = mx.sym.Dropout(data=hidden, p=dropout_rate)
+            if dropout > 0.:
+                hidden = mx.sym.Dropout(data=hidden, p=dropout)
 
             if direction == "forward":
                 hidden_all.append(hidden)
@@ -143,48 +155,26 @@ def gru_unroll(net, num_gru_layer, seq_len, num_hidden_gru_list, dropout_rate=0.
     return net
 
 
-def bi_gru_last_hiddens(net, num_gru_layer, seq_len, num_hidden_gru_list, dropout_rate=0., is_batchnorm=False):
+def bi_gru_unroll(net, num_gru_layer, seq_len, num_hidden_gru_list, dropout=0., is_batchnorm=False, is_bucketing=False):
     if num_gru_layer > 0:
         net_forward = gru_unroll(net=net,
                                  num_gru_layer=num_gru_layer,
                                  seq_len=seq_len,
                                  num_hidden_gru_list=num_hidden_gru_list,
-                                 dropout_rate=dropout_rate,
+                                 dropout=dropout,
                                  is_batchnorm=is_batchnorm,
                                  prefix="forward_",
-                                 direction="forward")
+                                 direction="forward",
+                                 is_bucketing=is_bucketing)
         net_backward = gru_unroll(net=net,
                                   num_gru_layer=num_gru_layer,
                                   seq_len=seq_len,
                                   num_hidden_gru_list=num_hidden_gru_list,
-                                  dropout_rate=dropout_rate,
+                                  dropout=dropout,
                                   is_batchnorm=is_batchnorm,
                                   prefix="backward_",
-                                  direction="backward")
-        hidden_all = []
-        hidden_all.append(mx.sym.Concat(*[net_forward[seq_len-1], net_backward[0]], dim=1))
-        net = hidden_all
-    return net
-
-
-def bi_gru_unroll(net, num_gru_layer, seq_len, num_hidden_gru_list, dropout_rate=0., is_batchnorm=False):
-    if num_gru_layer > 0:
-        net_forward = gru_unroll(net=net,
-                                 num_gru_layer=num_gru_layer,
-                                 seq_len=seq_len,
-                                 num_hidden_gru_list=num_hidden_gru_list,
-                                 dropout_rate=dropout_rate,
-                                 is_batchnorm=is_batchnorm,
-                                 prefix="forward_",
-                                 direction="forward")
-        net_backward = gru_unroll(net=net,
-                                  num_gru_layer=num_gru_layer,
-                                  seq_len=seq_len,
-                                  num_hidden_gru_list=num_hidden_gru_list,
-                                  dropout_rate=dropout_rate,
-                                  is_batchnorm=is_batchnorm,
-                                  prefix="backward_",
-                                  direction="backward")
+                                  direction="backward",
+                                  is_bucketing=is_bucketing)
         hidden_all = []
         for i in range(seq_len):
             hidden_all.append(mx.sym.Concat(*[net_forward[i], net_backward[i]], dim=1))
@@ -192,25 +182,27 @@ def bi_gru_unroll(net, num_gru_layer, seq_len, num_hidden_gru_list, dropout_rate
     return net
 
 
-def bi_gru_unroll_two_input_two_output(net1, net2, num_gru_layer, seq_len, num_hidden_gru_list, dropout_rate=0., is_batchnorm=False):
+def bi_gru_unroll_two_input_two_output(net1, net2, num_gru_layer, seq_len, num_hidden_gru_list, dropout=0.,
+                                       is_batchnorm=False, is_bucketing=False):
     if num_gru_layer > 0:
         net_forward = gru_unroll(net=net1,
                                  num_gru_layer=num_gru_layer,
                                  seq_len=seq_len,
                                  num_hidden_gru_list=num_hidden_gru_list,
-                                 dropout_rate=dropout_rate,
+                                 dropout=dropout,
                                  is_batchnorm=is_batchnorm,
                                  prefix="forward_",
-                                 direction="forward")
+                                 direction="forward",
+                                 is_bucketing=is_bucketing)
         net_backward = gru_unroll(net=net2,
                                   num_gru_layer=num_gru_layer,
                                   seq_len=seq_len,
                                   num_hidden_gru_list=num_hidden_gru_list,
-                                  dropout_rate=dropout_rate,
+                                  dropout=dropout,
                                   is_batchnorm=is_batchnorm,
                                   prefix="backward_",
-                                  direction="backward")
+                                  direction="backward",
+                                  is_bucketing=is_bucketing)
         return net_forward, net_backward
     else:
         return net1, net2
-
