+++ stt_layer_slice.py
